Simplified NamedNodeTest
