+++ src/main/scala/com/atomist/tree/pathexpression/NodeTest.scala
@@ -62,34 +62,3 @@ abstract class PredicatedNodeTest(name: String, predicate: Predicate) extends No
   */
 object All extends PredicatedNodeTest("All", TruePredicate)
 
-case class NamedNodeTest(name: String)
-  extends NodeTest {
-
-  override def follow(tn: TreeNode, axis: AxisSpecifier, ee: ExpressionEngine, typeRegistry: TypeRegistry): ExecutionResult = axis match {
-    case Child =>
-      tn match {
-        case ctn: ContainerTreeNode =>
-          val kids: List[TreeNode] =
-            ctn.childrenNamed(name).toList match {
-              case Nil =>
-                TreeNodeOperations.invokeMethodIfPresent[TreeNode](tn, name).
-                  map {
-                    case s: List[TreeNode@unchecked] =>
-                      s
-                    case t: TreeNode =>
-                      List(t)
-                    case x =>
-                      Nil
-                  }.getOrElse {
-                  val allKids = ctn.childNodes
-                  val filteredKids = allKids.filter(n => name.equals(n.nodeName))
-                  filteredKids.toList
-                }
-              case l => l
-            }
-          Right(kids)
-        case x => Left(s"Cannot find property [$name] on non-container tree node [$x]")
-      }
-  }
-
-}
