+++ src/test/resources/com/atomist/rug/kind/python3/Imports.ts
@@ -13,6 +13,12 @@ class Imports implements ProjectEditor {
       let eng: PathExpressionEngine = project.context().pathExpressionEngine()
 
       let i = 0
+
+      // eng.with<File>(project, "//File()[/PythonRawFile()//import_from()//FROM[@value='flask']]", n => {
+      //   console.log(`The file path is ${n.path()}`)
+      //   i++
+      // })
+
       eng.with<TreeNode>(project, "//File()/PythonRawFile()//import_stmt()", n => {
         console.log(`The node is ${n.value()}`)
         i++
